Added output and Database Access Program
If you want to run the program, the 'triviagame.db' file in the
TriviaMazeJDBC directory is the actual database, which is used by the
program. You may need to change your source path in eclipse to include
the file in the 'lib' folder, which is the jdbc that allows the program
to use sqlite, as well as the latest version of sqlite itself on your
computer. If you have any questions on all this just ask me, I already
struggled through it lol. I also just noticed some problems with the
program, but I have work so I can't fix any of the little bugs right
now. I'll do some more over the weekend to get them worked out.