Added Question Class and Screen Shots
Question class contains the question as a string, an index calue of
which is the correct one, and an arraylist of the answers themselves in
the form of strings. There are helper methods for printing, finding the
correct answer, etc. They are explained in the comments in the code and
should be useful in the game loop. Also added some screenshots of
functionality and the code in general, which are in the folder 'Database
Question Structure Screenshots'.