Added JUnit Tests and ScreenCaps
The JUnit tests had to have a different version of the code, so i just
made a new project with the revised code in it? They return strings
instead of print to the screen so the strings can be checked for
accuracy, maybe they could be like that in the final too I guess, let me
know which is easier to handle in the game loop. The revised code is in
the directory called "Database Access JUnit Tests", and the screen shots
are also in there. Any questions let me know!